[Cleanup] Create a clean folder `alex_suggested_tests`.
This folder was created as a clean approach compared to previous level folder which has now become too messy due to the fact that there are many test-sameeran-* tests added into it. This comes handy for targetting the tests which we aim as a plan when talking to Dr. Hall and Alex Z. The following test is marked currently XFAIL because maybe I was playing with it and changed it which I don't recall now, leaving it XFAIL as of now but fix it later