[Learning][TestCases] Added test cases while understanding-ir-structure.
I started with understanding the IR, I started writing the test cases by
referring the document understanding-ir-structure from mlir.llvm.org.
I started playing with the various concepts of modelling the IR. Doing
this activity gives me an understanding of reading the IR and various test
cases, even this will help in writing the tests which will be required for
writing the transform dialect test cases.