[Pass] Add a new test pass.
My observation is that the MLIR infrastructure has various ways of doing the same work
For example for adding a pass there are several ways, one is
Add a pass in test/ folder
other could be using the `PassRegistration` technique - similar to LLVM way
The above has become old way, so recent passes are migrated using the new way, I think
it's the unique pointer technique.

So this patch is a simpler one trying to add a new flag called `--dummy-pass` to mlir-opt.
The pass does nothing but just prints the Operation name.
The results of the test case are not printing the various other operations except ModuleOp.
I am not sure why, but this is a sample framework commit showing how to add a test case folder
pass, upcoming ones would try adding in different ways.

The command to run this would be
`mlir-opt --dummy-pass simple_test.mlir`